0.2.0
增加了内存池，连接池，提升了部分性能

+++ include/string_helper.h
@@ -11,7 +11,7 @@ public:
     static std::string GetRKeyValue(const std::string &source, const std::string left, const std::string right);
     static std::string Trim(const std::string &source, const std::string separators=" ");
     static std::string ToString(long long value);
-    static size_t Str2UInt64(std::string value);
+    static uint64_t Str2UInt64(std::string value);
 };
 
 #endif
